Bug fix
Changed storage type to float for storing values instead of long
Added input numbers to stack for later processing

+++ expr_part.h
@@ -2,7 +2,7 @@
 #define EXPR_PART_
 struct Part{
 	char operator;
-	long value;
+	float value;
 	struct Part *next;	
 };
 static void cleanup(struct Part *p){
